[B] Fix globalState race condition in randomData
Previously, when random data option was true GalaxiesSelector fetched
random data and then immediately created an answer in order to persist
the data across pages.  That was causing a race condition between the
fetch and GlobalState updating when visiting the page for the first
with nothing previously saved in localStorage.  This update punts the
answer creation to DidUpdate avoiding since the answer creation is not
vital until after you click thorugh to a different page.  THis avoids
the race altogether