[F] OrbitViewer details table fits inside component footprint
Previously the details table was so big it had to
go under the fold and below the OrbitViewer. This
update nests the details table within the
OrbitViewer and provides a toggle button to
reveal/hide the details. Toggle button is disabled
if no object details.

Updated details table includes all relevant object
details